+++ benight-admin/src/app/navigator/navigator-routing.module.ts
@@ -4,8 +4,20 @@ import { NavigatorPage } from './navigator.page'
 
 const routes: Routes = [{
   path:'',
-  component:NavigatorPage
-}]
+  component: NavigatorPage,
+  children: [
+    { path: '', redirectTo: 'eventos-handler', pathMatch: 'full' },
+    { path: 'gestion', loadChildren: './gestion/gestion.module#GestionPageModule' },
+    { path: 'chat', loadChildren: './chat/chat.module#ChatPageModule' },
+    { path: 'notificaciones', loadChildren: './notificaciones/notificaciones.module#NotificacionesPageModule' },
+    { path: 'general', loadChildren: './general/general.module#GeneralPageModule' },
+    { path: '**', redirectTo: '', pathMatch: 'full' }    
+  ]
+},
+{ path: '**', redirectTo: '', pathMatch: 'full' },
+  
+
+]
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],