+++ benight-employee/src/app/navigator/navigator.page.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core'
+import { SharedDataService } from '@bn8-services/shared-data.service'
 
 @Component({
   selector: 'navigator',
@@ -7,7 +8,15 @@ import { Component, OnInit } from '@angular/core'
 })
 export class NavigatorPage implements OnInit {
 
-  constructor() { }
+  mainTabs = [
+    {href:'/navigator/eventos-handler',icon:'play',title:'Eventos'},
+    {href:'/navigator/entradas',icon:'list',title:'Entradas'},
+    {href:'/navigator/chat',icon:'list',title:'Chat'},
+    {href:'/navigator/notificaciones',icon:'book',title:'Notificaciones'},
+    {href:'/navigator/perfil',icon:'people',title:'Perfil'}
+  ] 
+
+  constructor(public sd: SharedDataService) { }
 
   ngOnInit() {
   }