+++ benight/src/app/navigator/navigator.module.ts
@@ -1,4 +1,3 @@
-
 import { NgModule } from '@angular/core'
 
 import { NavigatorPage } from './navigator.page'