+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 ==========
 
 
+4th commit (2018-10-4, 22:00)
+--------------------
+Benight-Admin:
++ Rudimentary navigation added
+
+Benight-Employee:
++ Rudimentary navigation added
+
+
 3rd commit (2018-10-4, 21:00)
 --------------------
-enight-Admin:
+Benight-Admin:
 + Code Reorder preparing the navigation
 
 Benight-Employee: